pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-mcp-telegram"
version = "0.1.0"
description = "MCP Server for Codex CLI with Telegram integration for remote access"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "python-telegram-bot>=21.0.0",
]

[project.scripts]
codex-mcp-telegram = "codex_mcp_server.server:main"
codex-mcp-server = "codex_mcp_server.server:main"  # Backward compatibility alias

[tool.setuptools]
packages = ["codex_mcp_server"]
